Ensure same QR code is retained during 2FA setup
When setting up app based 2FA a QR code is displayed to the user to
scan into an authenticator app. If the user enters an invalid code,
an entirely new QR code would be generated which may confuse the
user.

This PR fixes this and ensures the same QR code is displayed to the
user if the code entered is wrong.

Trello: https://trello.com/c/SmJsuBoG/378-fix-issue-with-qr-code